blossom system update, blossom level status, mobile

+++ deletepost.php
@@ -18,31 +18,16 @@ $blossom_progress = $connect->query("SELECT blossom_progress FROM users WHERE id
 $blossom = $blossom_level + $blossom_progress / 100;
 $blossom -= 0.1;
 
-$user_level = intval($blossom);
-$user_progress = ($blossom - $user_level) * 100;
-
-$connect->query("UPDATE users SET blossom_level = $user_level WHERE id = $user_id");
-$connect->query("UPDATE users SET blossom_progress = $user_progress WHERE id = $user_id");
-
 $other_users_id_comments = $connect->query("SELECT user_id FROM comments WHERE post_id = $post_id");
 
 if ($other_users_id_comments->num_rows > 0) {
     while ($row = $other_users_id_comments->fetch_assoc()) {
         $other_id = $row['user_id'];
-        $other_blossom_level = $connect->query("SELECT blossom_level FROM users WHERE id = $other_id")->fetch_assoc()['blossom_level'];
-        $other_blossom_progress = $connect->query("SELECT blossom_progress FROM users WHERE id = $other_id")->fetch_assoc()['blossom_progress'];
-        $other_blossom = $other_blossom_level + $other_blossom_progress / 100;
-        $other_blossom -= 0.03;
+        $blossom -= 0.04;
         
         if ($other_id == $user_id) {
-            $other_blossom -= 0.04;
+            $blossom -= 0.03;
         }
-
-        $other_user_level = intval($other_blossom);
-        $other_user_progress = ($other_blossom - $other_user_level) * 100;
-
-        $connect->query("UPDATE users SET blossom_level = $other_user_level WHERE id = $other_id");
-        $connect->query("UPDATE users SET blossom_progress = $other_user_progress WHERE id = $other_id");
     }
 }
 
@@ -51,26 +36,23 @@ $other_users_id_likes = $connect->query("SELECT user_id FROM likes_on_posts WHER
 if ($other_users_id_likes->num_rows > 0) {
     while ($row = $other_users_id_likes->fetch_assoc()) {
         $other_id = $row['user_id'];
-        $other_blossom_level = $connect->query("SELECT blossom_level FROM users WHERE id = $other_id")->fetch_assoc()['blossom_level'];
-        $other_blossom_progress = $connect->query("SELECT blossom_progress FROM users WHERE id = $other_id")->fetch_assoc()['blossom_progress'];
-        $other_blossom = $other_blossom_level + $other_blossom_progress / 100;
-        $other_blossom -= 0.02;
+        $blossom -= 0.03;
         
         if ($other_id == $user_id) {
-            $other_blossom -= 0.03;
+            $blossom -= 0.02;
         }
-
-        $other_user_level = intval($other_blossom);
-        $other_user_progress = ($other_blossom - $other_user_level) * 100;
-
-        $connect->query("UPDATE users SET blossom_level = $other_user_level WHERE id = $other_id");
-        $connect->query("UPDATE users SET blossom_progress = $other_user_progress WHERE id = $other_id");
     }
 }
 
+$user_level = intval($blossom);
+$user_progress = ($blossom - $user_level) * 100;
+
+$connect->query("UPDATE users SET blossom_level = $user_level WHERE id = $user_id");
+$connect->query("UPDATE users SET blossom_progress = $user_progress WHERE id = $user_id");
+
 $connect->query("DELETE FROM posts WHERE id = $post_id AND user_id = $user_id");
-$connect->query("DELETE FROM comments WHERE post_id = $post_id");
-$connect->query("DELETE FROM likes_on_posts WHERE post_id = $post_id");
+$connect->query("DELETE FROM comments WHERE post_id = $post_id AND user_id = $user_id");
+$connect->query("DELETE FROM likes_on_posts WHERE post_id = $post_id AND user_id = $user_id");
 
 if ($source == 'profile') {
     header('Location: ./profile');
